pyproject.toml
[project]
name = "propertyai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]==0.116.1",
    "logger>=1.4",
    "openai>=1.102.0",
    "pydantic-settings>=2.10.1",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "pyyaml==6.0.2",
    "qdrant-client>=1.15.1",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.12.10",
]
